move build_root_params_from_runconfig() to be a classmethod in RootParamGroup
